pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-openai-assistant"
version = "0.7.8"
dependencies = ["openai", "markdown"]
readme = "README.md"
authors = [{name = "Jean-Luc Vanhulst", email = "jl@valor.vc"}]
maintainers = [{name = "Jean-Luc Vanhulst", email = "jl@valor.vc"}]
classifiers = ["License :: OSI Approved :: MIT License"]
urls = {Homepage = "https://github.com/jlvanhulst/django-openai"}
keywords = ["django", "celery", "openai", "assistants"]
description = "Django OpenAI Assistant"

[project.optional-dependencies]
dev = [
    "black>=22.3.0",
    "flake8>=4.0.1",
    "isort",
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "pytest-celery>=0.0.0",
    "pytest-redis>=3.0.0",
    "pytest-mock",
    "pytest-asyncio"
]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist"]

[tool.hatch.build]
exclude = ["~/*",".venv",".vscode","build"]


